refactor: move the embedded JavaScript into files a linter can read
`lib.rs` carried ~2,500 lines of JavaScript inside Rust raw strings, where no
TypeScript or lint tooling could see it. That blindness has cost real defects:
a probe that queried a deleted renderer's selectors for weeks — its own contract
test's header says "the same accident happened three times" — and the 176 dead
literal tokens removed in the previous commit, every one of which `tsc` would
have flagged in a second.

The four non-templated blocks move to `src-tauri/src/webview_verify/*.js`, loaded
with `include_str!`. The three `format!` templates stay inline: converting them
needs a placeholder scheme, which is a behaviour-change risk this move should not
carry.

Nothing about the built binary changes. Each extracted file was compared byte for
byte against the string it replaced — 127, 100, 250 and 2,028 lines, all four
identical, verified independently rather than taken on the extractor's word.
`lib.rs` drops from 7,026 to 4,530 lines.

It paid for itself immediately. With the probe in a real `.js` file, `eslint`
found `markerSummary` — a helper defined and never called, invisible for as long
as it lived in a string. It is gone.

The four gates that parsed the JavaScript out of the Rust source now read the
files directly, which also simplifies them: `probe-selectors.ts` existed because
"the probes are JS inside a Rust raw string, so static parsing is impossible",
and that is no longer true.
